+++ src/prognostic_equations/forcing/external_forcing.jl
@@ -318,7 +318,7 @@ function external_forcing_tendency!(
     # horizontal advection, vertical fluctuation, nudging, subsidence (need to add),
     (; params) = p
     thermo_params = CAP.thermodynamics_params(params)
-    (; ᶜts) = p.precomputed
+    (; ᶜT) = p.precomputed
     (;
         ᶜdTdt_fluc,
         ᶜdqtdt_fluc,
@@ -338,13 +338,12 @@ function external_forcing_tendency!(
     @. ᶜuₕ_nudge = C12(Geometry.UVVector(ᶜu_nudge, ᶜv_nudge), ᶜlg)
     @. Yₜ.c.uₕ -= (Y.c.uₕ - ᶜuₕ_nudge) * ᶜinv_τ_wind
 
-    ᶜe_tot = @. lazy(specific(Y.c.ρe_tot, Y.c.ρ))
-    ᶜh_tot = @. lazy(TD.total_specific_enthalpy(thermo_params, ᶜts, ᶜe_tot))
+    (; ᶜh_tot, ᶜq_tot_safe, ᶜq_liq_rai, ᶜq_ice_sno) = p.precomputed
     # nudging tendency
     ᶜdTdt_nudging = p.scratch.ᶜtemp_scalar
     ᶜdqtdt_nudging = p.scratch.ᶜtemp_scalar_2
     @. ᶜdTdt_nudging =
-        -(TD.air_temperature(thermo_params, ᶜts) - ᶜT_nudge) * ᶜinv_τ_scalar
+        -(ᶜT - ᶜT_nudge) * ᶜinv_τ_scalar
     @. ᶜdqtdt_nudging =
         -(specific(Y.c.ρq_tot, Y.c.ρ) - ᶜqt_nudge) * ᶜinv_τ_scalar
 
@@ -360,9 +359,9 @@ function external_forcing_tendency!(
     # total energy
     @. Yₜ.c.ρe_tot +=
         Y.c.ρ * (
-            TD.cv_m(thermo_params, ᶜts) * ᶜdTdt_sum +
+            TD.cv_m(thermo_params, ᶜq_tot_safe, ᶜq_liq_rai, ᶜq_ice_sno) * ᶜdTdt_sum +
             (
-                cv_v * (TD.air_temperature(thermo_params, ᶜts) - T_0) + Lv_0 -
+                cv_v * (ᶜT - T_0) + Lv_0 -
                 R_v * T_0
             ) * ᶜdqtdt_sum
         )
@@ -561,7 +560,7 @@ function external_forcing_tendency!(Yₜ, Y, p, t, ::ISDACForcing)
     FT = Spaces.undertype(axes(Y.c))
     (; params) = p
     thermo_params = CAP.thermodynamics_params(params)
-    (; ᶜts, ᶜp) = p.precomputed
+    (; ᶜp, ᶜT) = p.precomputed
 
     ᶜinv_τ_scalar = APL.ISDAC_inv_τ_scalar(FT)  # s⁻¹
     ᶜinv_τ_wind = APL.ISDAC_inv_τ_wind(FT)  # s⁻¹
@@ -588,7 +587,7 @@ function external_forcing_tendency!(Yₜ, Y, p, t, ::ISDACForcing)
     ᶜdTdt_nudging = p.scratch.ᶜtemp_scalar
     ᶜdqtdt_nudging = p.scratch.ᶜtemp_scalar_2
     @. ᶜdTdt_nudging =
-        -(TD.air_temperature(thermo_params, ᶜts) - ta_ISDAC(ᶜp, ᶜz)) *
+        -(ᶜT - ta_ISDAC(ᶜp, ᶜz)) *
         ᶜinv_τ_scalar(ᶜz)
     @. ᶜdqtdt_nudging =
         -(specific(Y.c.ρq_tot, Y.c.ρ) - q_tot(ᶜz)) * ᶜinv_τ_scalar(ᶜz)
@@ -598,11 +597,12 @@ function external_forcing_tendency!(Yₜ, Y, p, t, ::ISDACForcing)
     cv_v = TD.Parameters.cv_v(thermo_params)
     R_v = TD.Parameters.R_v(thermo_params)
     # total energy
+    (; ᶜq_tot_safe, ᶜq_liq_rai, ᶜq_ice_sno) = p.precomputed
     @. Yₜ.c.ρe_tot +=
         Y.c.ρ * (
-            TD.cv_m(thermo_params, ᶜts) * ᶜdTdt_nudging +
+            TD.cv_m(thermo_params, ᶜq_tot_safe, ᶜq_liq_rai, ᶜq_ice_sno) * ᶜdTdt_nudging +
             (
-                cv_v * (TD.air_temperature(thermo_params, ᶜts) - T_0) + Lv_0 -
+                cv_v * (ᶜT - T_0) + Lv_0 -
                 R_v * T_0
             ) * ᶜdqtdt_nudging
         )